feat: fractional baths on properties (1.5 / 2.5) end to end
properties.baths was INTEGER, so half-bath values were rejected at the
DB level. Switch to NUMERIC(4,1) (migration 0062), accept decimals in
both opportunity forms, and relax the baths inputs to 0.5 steps.

+++ FrameworkPlanner/server/crm/import-export.ts
@@ -74,7 +74,7 @@ const opportunityFields: FieldDef[] = [
   { key: "state", label: "State", required: true, type: "string" },
   { key: "zipCode", label: "Zip Code", required: true, type: "string" },
   { key: "beds", label: "Beds", type: "int" },
-  { key: "baths", label: "Baths", type: "int" },
+  { key: "baths", label: "Baths", type: "decimal" },
   { key: "sqft", label: "Square Feet", type: "int" },
   { key: "price", label: "Price", type: "decimal" },
   { key: "status", label: "Status", type: "string" },
